Squashed commit of the following:

commit d7633e38dde62c675be84fe6fb2427719bb6e36f
Merge: df1e381 2f8414e
Author: Michał Grygierzec <michal.grygierzec@relativity.com>
Date:   Mon Dec 18 15:01:16 2017 +0100

    Merge branch 'develop' into REL-170226_int_to_long

commit df1e381f537f51b4cca83872416985923fe1e403
Author: Michał Grygierzec <michal.grygierzec@relativity.com>
Date:   Mon Dec 18 14:29:48 2017 +0100

    Fix image file statistics unit tests

commit 48d5dd229701ca43c4de40cf2460c588d3653d47
Author: Michał Grygierzec <michal.grygierzec@relativity.com>
Date:   Mon Dec 18 13:36:28 2017 +0100

    Refactor also batching mechanism to use long instead of int.

commit 8bf68e31b5cc7e3b3f28926cfc8d6522185133dd
Author: Michał Grygierzec <michal.grygierzec@relativity.com>
Date:   Mon Dec 18 13:35:41 2017 +0100

    Change int to long in document statistics calculations.

    It is used by Total and FileSize statistics for Documents, Images and Native files. Please note that total documents count mechanism cannot be changed without modifications outside RIP (kCura.Relativity.Client.dll). Now we just case implicitly from int to long inside total count methods.

Merge pull request #594 in IN/integrationpoints from REL-201888-r1-incident-integration-points to release-9.6-april

Squashed commit of the following:

commit fac1f35b403b72b0a37c011e930d7a24d2ebe60c
Author: Michał Grygierzec <michal.grygierzec@relativity.com>
Date:   Tue Feb 20 14:56:15 2018 +0100

    Fix Unit Tests for NativeFileSizeStatistics

commit e6c1519d55a630adbb35ad6229ca8b55cc06047d
Author: Michał Grygierzec <michal.grygierzec@relativity.com>
Date:   Tue Feb 20 14:54:19 2018 +0100

    [REL-201888] Fix Integration points error Job status failed

(cherry picked from commit adb81c1398134bfbbdf96f6ceac69bdf118d5a63)

+++ Source/kCura.IntegrationPoints.Agent/Tasks/ExportManager.cs
@@ -86,7 +86,7 @@ namespace kCura.IntegrationPoints.Agent.Tasks
 		    yield break;
 		}
 
-        public override int BatchTask(Job job, IEnumerable<string> batchIDs)
+        public override long BatchTask(Job job, IEnumerable<string> batchIDs)
 		{
 		    LogBatchTaskStart(job, batchIDs);
             var integrationPoint = IntegrationPointService.GetRdo(job.RelatedObjectArtifactID);
@@ -100,7 +100,7 @@ namespace kCura.IntegrationPoints.Agent.Tasks
 			ExportUsingSavedSearchSettings sourceSettings = Serializer.Deserialize<ExportUsingSavedSearchSettings>(integrationPoint.SourceConfiguration);
 			DestinationConfiguration destinationConfiguration = JsonConvert.DeserializeObject<DestinationConfiguration>(integrationPoint.DestinationConfiguration);
 
-			int totalCount = GetTotalExportItemsCount(sourceSettings, destinationConfiguration, job);
+			long totalCount = GetTotalExportItemsCount(sourceSettings, destinationConfiguration, job);
 
 			if (totalCount > 0)
 			{
@@ -111,12 +111,12 @@ namespace kCura.IntegrationPoints.Agent.Tasks
             return totalCount;
 		}
 
-	    private int GetTotalExportItemsCount(ExportUsingSavedSearchSettings settings, DestinationConfiguration destinationConfiguration, Job job)
+	    private long GetTotalExportItemsCount(ExportUsingSavedSearchSettings settings, DestinationConfiguration destinationConfiguration, Job job)
 		{
 			try
 			{
                 LogGetTotalExportItemsCountStart(job);
-                int count = _exportInitProcessService.CalculateDocumentCountToTransfer(settings, destinationConfiguration.ArtifactTypeId);
+                long count = _exportInitProcessService.CalculateDocumentCountToTransfer(settings, destinationConfiguration.ArtifactTypeId);
 			    LogGetTotalExportItemsCountSuccesfulEnd(job, count);
 			    return count;
 			}
@@ -152,7 +152,7 @@ namespace kCura.IntegrationPoints.Agent.Tasks
 	            batchIDs);
 	    }
 
-	    private void LogBatchTaskSuccesfulEnd(Job job, int totalCount)
+	    private void LogBatchTaskSuccesfulEnd(Job job, long totalCount)
 	    {
 	        _logger.LogInformation("Finished batch task in Export Manager for job: {JobId}, totalCount: {totalCount}",
 	            job.JobId, totalCount);
@@ -163,7 +163,7 @@ namespace kCura.IntegrationPoints.Agent.Tasks
 	        _logger.LogInformation("Trying to get total export items count for job: {JobId}", job.JobId);
 	    }
 
-	    private void LogGetTotalExportItemsCountSuccesfulEnd(Job job, int count)
+	    private void LogGetTotalExportItemsCountSuccesfulEnd(Job job, long count)
 	    {
 	        _logger.LogInformation("Retrieved total export items count for job: {JobId}, count: {count}", job.JobId, count);
 	    }
