+++ Source/kCura.IntegrationPoints.Services.Tests.Integration/StatisticsManager/TestCase/GetNativesFileSizeForSavedSearch.cs
@@ -7,7 +7,7 @@ namespace kCura.IntegrationPoints.Services.Tests.Integration.StatisticsManager.T
 	{
 		public void Execute(ITestHelper helper, int workspaceArtifactId, TestCaseSettings testCaseSettings)
 		{
-			int total;
+			long total;
 			using (var statisticsManager = helper.CreateAdminProxy<IStatisticsManager>())
 			{
 				total = statisticsManager.GetNativesFileSizeForSavedSearchAsync(workspaceArtifactId, testCaseSettings.SavedSearchId).Result;
