+++ Source/kCura.IntegrationPoints.Services/StatisticsManager.cs
@@ -23,97 +23,97 @@ namespace kCura.IntegrationPoints.Services
 		{
 		}
 
-		public Task<int> GetDocumentsTotalForSavedSearchAsync(int workspaceArtifactId, int savedSearchId)
+		public Task<long> GetDocumentsTotalForSavedSearchAsync(int workspaceArtifactId, int savedSearchId)
 		{
 			return ForSavedSearch<IDocumentTotalStatistics>(workspaceArtifactId, savedSearchId);
 		}
 
-		public Task<int> GetNativesTotalForSavedSearchAsync(int workspaceArtifactId, int savedSearchId)
+		public Task<long> GetNativesTotalForSavedSearchAsync(int workspaceArtifactId, int savedSearchId)
 		{
 			return ForSavedSearch<INativeTotalStatistics>(workspaceArtifactId, savedSearchId);
 		}
 
-		public Task<int> GetImagesTotalForSavedSearchAsync(int workspaceArtifactId, int savedSearchId)
+		public Task<long> GetImagesTotalForSavedSearchAsync(int workspaceArtifactId, int savedSearchId)
 		{
 			return ForSavedSearch<IImageTotalStatistics>(workspaceArtifactId, savedSearchId);
 		}
 
-		public Task<int> GetImagesFileSizeForSavedSearchAsync(int workspaceArtifactId, int savedSearchId)
+		public Task<long> GetImagesFileSizeForSavedSearchAsync(int workspaceArtifactId, int savedSearchId)
 		{
 			return ForSavedSearch<IImageFileSizeStatistics>(workspaceArtifactId, savedSearchId);
 		}
 
-		public Task<int> GetNativesFileSizeForSavedSearchAsync(int workspaceArtifactId, int savedSearchId)
+		public Task<long> GetNativesFileSizeForSavedSearchAsync(int workspaceArtifactId, int savedSearchId)
 		{
 			return ForSavedSearch<INativeFileSizeStatistics>(workspaceArtifactId, savedSearchId);
 		}
 
-		public Task<int> GetDocumentsTotalForProductionAsync(int workspaceArtifactId, int productionSetId)
+		public Task<long> GetDocumentsTotalForProductionAsync(int workspaceArtifactId, int productionSetId)
 		{
 			return ForProduction<IDocumentTotalStatistics>(workspaceArtifactId, productionSetId);
 		}
 
-		public Task<int> GetNativesTotalForProductionAsync(int workspaceArtifactId, int productionSetId)
+		public Task<long> GetNativesTotalForProductionAsync(int workspaceArtifactId, int productionSetId)
 		{
 			return ForProduction<INativeTotalStatistics>(workspaceArtifactId, productionSetId);
 		}
 
-		public Task<int> GetImagesTotalForProductionAsync(int workspaceArtifactId, int productionSetId)
+		public Task<long> GetImagesTotalForProductionAsync(int workspaceArtifactId, int productionSetId)
 		{
 			return ForProduction<IImageTotalStatistics>(workspaceArtifactId, productionSetId);
 		}
 
-		public Task<int> GetImagesFileSizeForProductionAsync(int workspaceArtifactId, int productionSetId)
+		public Task<long> GetImagesFileSizeForProductionAsync(int workspaceArtifactId, int productionSetId)
 		{
 			return ForProduction<IImageFileSizeStatistics>(workspaceArtifactId, productionSetId);
 		}
 
-		public Task<int> GetNativesFileSizeForProductionAsync(int workspaceArtifactId, int productionSetId)
+		public Task<long> GetNativesFileSizeForProductionAsync(int workspaceArtifactId, int productionSetId)
 		{
 			return ForProduction<INativeFileSizeStatistics>(workspaceArtifactId, productionSetId);
 		}
 
-		public Task<int> GetDocumentsTotalForFolderAsync(int workspaceArtifactId, int folderId, int viewId, bool includeSubFoldersTotals)
+		public Task<long> GetDocumentsTotalForFolderAsync(int workspaceArtifactId, int folderId, int viewId, bool includeSubFoldersTotals)
 		{
 			return ForFolder<IDocumentTotalStatistics>(workspaceArtifactId, folderId, viewId, includeSubFoldersTotals);
 		}
 
-		public Task<int> GetNativesTotalForFolderAsync(int workspaceArtifactId, int folderId, int viewId, bool includeSubFoldersTotals)
+		public Task<long> GetNativesTotalForFolderAsync(int workspaceArtifactId, int folderId, int viewId, bool includeSubFoldersTotals)
 		{
 			return ForFolder<INativeTotalStatistics>(workspaceArtifactId, folderId, viewId, includeSubFoldersTotals);
 		}
 
-		public Task<int> GetImagesTotalForFolderAsync(int workspaceArtifactId, int folderId, int viewId, bool includeSubFoldersTotals)
+		public Task<long> GetImagesTotalForFolderAsync(int workspaceArtifactId, int folderId, int viewId, bool includeSubFoldersTotals)
 		{
 			return ForFolder<IImageTotalStatistics>(workspaceArtifactId, folderId, viewId, includeSubFoldersTotals);
 		}
 
-		public Task<int> GetImagesFileSizeForFolderAsync(int workspaceArtifactId, int folderId, int viewId, bool includeSubFoldersTotals)
+		public Task<long> GetImagesFileSizeForFolderAsync(int workspaceArtifactId, int folderId, int viewId, bool includeSubFoldersTotals)
 		{
 			return ForFolder<IImageFileSizeStatistics>(workspaceArtifactId, folderId, viewId, includeSubFoldersTotals);
 		}
 
-		public Task<int> GetNativesFileSizeForFolderAsync(int workspaceArtifactId, int folderId, int viewId, bool includeSubFoldersTotals)
+		public Task<long> GetNativesFileSizeForFolderAsync(int workspaceArtifactId, int folderId, int viewId, bool includeSubFoldersTotals)
 		{
 			return ForFolder<INativeFileSizeStatistics>(workspaceArtifactId, folderId, viewId, includeSubFoldersTotals);
 		}
 
-		private Task<int> ForSavedSearch<T>(int workspaceArtifactId, int savedSearchId) where T : IDocumentStatistics
+		private Task<long> ForSavedSearch<T>(int workspaceArtifactId, int savedSearchId) where T : IDocumentStatistics
 		{
 			return Execute<T>(statistics => statistics.ForSavedSearch(workspaceArtifactId, savedSearchId), workspaceArtifactId);
 		}
 
-		private Task<int> ForProduction<T>(int workspaceArtifactId, int productionId) where T : IDocumentStatistics
+		private Task<long> ForProduction<T>(int workspaceArtifactId, int productionId) where T : IDocumentStatistics
 		{
 			return Execute<T>(statistics => statistics.ForProduction(workspaceArtifactId, productionId), workspaceArtifactId);
 		}
 
-		private Task<int> ForFolder<T>(int workspaceArtifactId, int folderId, int viewId, bool includeSubFoldersTotals) where T : IDocumentStatistics
+		private Task<long> ForFolder<T>(int workspaceArtifactId, int folderId, int viewId, bool includeSubFoldersTotals) where T : IDocumentStatistics
 		{
 			return Execute<T>(statistics => statistics.ForFolder(workspaceArtifactId, folderId, viewId, includeSubFoldersTotals), workspaceArtifactId);
 		}
 
-		private Task<int> Execute<T>(Func<IDocumentStatistics, int> action, int workspaceArtifactId) where T : IDocumentStatistics
+		private Task<long> Execute<T>(Func<IDocumentStatistics, long> action, int workspaceArtifactId) where T : IDocumentStatistics
 		{
 			CheckPermissions($"{nameof(StatisticsManager)} - {nameof(T)}", workspaceArtifactId);
 			try
